小修复

+++ sh/alipan-autoupload.py
@@ -12,13 +12,13 @@ from pathlib import Path
 from aligo import Aligo
 
 # 上传到阿里云盘
-def upload_to_alipan(file_path: Path):
+def upload_to_alipan(net_path: Path, file_path: Path):
     ali = Aligo("calllivecn")
-    remote_folder = ali.get_folder_by_path(file_path.parent.as_posix())
+    remote_folder = ali.get_folder_by_path(net_path.parent.as_posix())
 
     if remote_folder is None:
-        print(f"远程目录不存在: {file_path.parent}, 现在创建")
-        remote_folder = ali.create_folder(file_path.parent.as_posix())
+        print(f"远程目录不存在: {net_path.parent}, 现在创建")
+        remote_folder = ali.create_folder(net_path.parent.as_posix())
 
     ali.upload_file(file_path, remote_folder.file_id)
 
@@ -27,7 +27,7 @@ def monitor_dir(dir_path: Path, file_queue: queue.Queue):
     cmd = [
         "inotifywait", "-mrq",
         "--format", r"%w%f",
-        "-e", "close_write",
+        "-e", "move,close_write",
         dir_path
     ]
     with subprocess.Popen(cmd, stdout=subprocess.PIPE, text=True) as proc:
@@ -48,9 +48,10 @@ def clear_empty_dirs(root_dir: Path):
         for dirpath, dirnames, filenames in root_dir.walk(top_down=False):
             if not dirnames and not filenames:
                 try:
-                    dirpath.rmdir()
-                    empty_found = True
-                    print(f"删除空目录: {dirpath}")
+                    if root_dir != dirpath:
+                        dirpath.rmdir()
+                        empty_found = True
+                        print(f"删除空目录: {dirpath}")
                 except OSError as e:
                     print(f"无法删除目录 {dirpath}: {e}")
 
@@ -81,9 +82,14 @@ def main():
         print(f"处理文件: {file_path}")
 
         # 在这里添加你的处理逻辑
-        upload_to_alipan(file_path)
-        file_path.unlink(missing_ok=True)
-        print(f"上传完成，已删除本地文件: {file_path}")
+        net_path = file_path.relative_to(watch_dir)
+
+        if file_path.is_file():
+            upload_to_alipan(net_path, file_path)
+            file_path.unlink(missing_ok=True)
+            print(f"上传完成，已删除本地文件: {file_path}")
+        else:
+            print(f"本地文件不存在: {file_path}")
         file_queue.task_done()
